Convert submission fields to a preference.
Moves the hard-coded status, grade, grader, modified, and extension fields
into a preference, and add other grading fields:
 - gradevalue - actual grade value, separate from maximum grade.
 - grademax - maximum grade
 - created - submission creation date
 - latenessincext - how late the submission was, including any extensions

The fields gradevalue and grademax are intended as an alternative to
grade, suitable for further processing outside of Moodle, via
spreadsheet.

The default choices are set to the previous hard-coded fields.

NB: This has changed the field order in the columns.