Add a placholder function for debugging.
Trying to track down which field is getting an extra column can be
pretty frustrating if there's no data that can be used to identify
the field - when it's just the '-' empty value placeholder.

Change false to true in the placeholder() function and it will put field
ids in for many (not all) placeholders.